refactor: Migrating pages and routing hooks
